move custom_warning from tools to util
